fixes some compiler warnings

+++ GHStringHashTable.h
@@ -10,5 +10,5 @@ public:
 	virtual ~GHStringHashTable(void) {}
 	
     // if len == 0 use the full length of the string
-	virtual GHIdentifier generateHash(const char* str, int len=0) const = 0;
+	virtual GHIdentifier generateHash(const char* str, size_t len=0) const = 0;
 };
